fix: 타입 지정 모듈 수정
