+++ src/lib/types/Profile.ts
@@ -19,13 +19,3 @@ export type Profile = {
   updatedAt: string;
   name: string;
 };
-
-export type Pagination = {
-  page: number;
-  pageSize: number;
-};
-
-export type ProfileList = {
-  totalCount: number;
-  list: Profile[];
-};
